+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/SoraMediaChannel.kt
@@ -217,9 +217,9 @@ class SoraMediaChannel(
         }
 
         override fun onAddRemoteStream(ms: MediaStream) {
-            SoraLogger.d(TAG, "[channel:$role] @peer:onAddRemoteStream:${ms.label()}")
-            if (mediaOption.upstreamIsRequired && clientId != null && ms.label() == clientId) {
-                SoraLogger.d(TAG, "[channel:$role] this stream is mine, ignore: ${ms.label()}")
+            SoraLogger.d(TAG, "[channel:$role] @peer:onAddRemoteStream:${ms.id}")
+            if (mediaOption.upstreamIsRequired && clientId != null && ms.id == clientId) {
+                SoraLogger.d(TAG, "[channel:$role] this stream is mine, ignore: ${ms.id}")
                 return
             }
             listener?.onAddRemoteStream(this@SoraMediaChannel, ms)